+++ ConsoleRPG_Team/.vs/ConsoleRPG_Team/v17/DocumentLayout.backup.json
@@ -42,9 +42,9 @@
               "Title": "Program.cs",
               "DocumentMoniker": "C:\\Users\\kheam\\OneDrive\\\uBB38\uC11C\\GitHub\\ConsoleRPG_TeamProject\\ConsoleRPG_Team\\Program.cs",
               "RelativeDocumentMoniker": "Program.cs",
-              "ToolTip": "C:\\Users\\kheam\\OneDrive\\\uBB38\uC11C\\GitHub\\ConsoleRPG_TeamProject\\ConsoleRPG_Team\\Program.cs",
-              "RelativeToolTip": "Program.cs",
-              "ViewState": "AgIAAE0AAAAAAAAAAAAYwGIAAAALAAAAAAAAAA==",
+              "ToolTip": "C:\\Users\\kheam\\OneDrive\\\uBB38\uC11C\\GitHub\\ConsoleRPG_TeamProject\\ConsoleRPG_Team\\Program.cs*",
+              "RelativeToolTip": "Program.cs*",
+              "ViewState": "AgIAAAAAAAAAAAAAAAAAAA4AAAAAAAAAAAAAAA==",
               "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000738|",
               "WhenOpened": "2025-07-14T08:20:52.809Z",
               "EditorCaption": ""
